Remove internal funcs, rename _mlflow.py
